Implemented tree view recursion now should work properly

+++ EtnaPOS.DAL/Models/Artikal.cs
@@ -3,7 +3,7 @@
     public class Artikal : BaseEntity
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int KategorijaArtiklaId { get; set; }
         public KategorijaArtikla KategorijaArtikla { get; set; }
         public bool IsActive { get; set; }
@@ -14,7 +14,7 @@
         public Artikal(string name, double price, int kategorijaId)
         {
             Name = name;
-            Price = price;
+            Price = (Decimal)price;
             KategorijaArtiklaId = kategorijaId;
             IsActive = true;
         }
@@ -22,7 +22,7 @@
         {
             Id = id;
             Name = name;
-            Price = price;
+            Price = (Decimal)price;
             KategorijaArtiklaId = kategorijaId;
             IsActive = true;
         }
